+++ src/modules/project_status/index.tsx
@@ -1,2 +1,3 @@
-export {DetailProjectStatus} from "./view/detail_project_status"
-export {ListProjectStatus} from "./view/list_project_status"
+export { DetailProjectStatus } from "./view/detail_project_status"
+export { ListProjectStatus } from "./view/list_project_status"
+export { ActivitiesAnalyticProjects } from "./view/list_project_activities_analytic"
